requirements.txt
llama-index
sentence-transformers
fastapi
uvicorn
python-telegram-bot
pyyaml
httpx 